Export ip validation and use it in add_ip & tests
